requirements.txt
alembic==1.8.1
autopep8==2.0.0
bcrypt==4.0.1
blinker==1.5
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-Admin==1.6.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
Mako==1.2.4
MarkupSafe==2.1.1
mysqlclient==2.1.1
Pillow==9.4.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
SQLAlchemy==1.4.44
tomli==2.0.1
typing_extensions==4.4.0
Werkzeug==2.2.2
WTForms==3.0.1
WTForms-SQLAlchemy==0.3
zipp==3.11.0
